test(cathedral-5): v131-aware v126 shape pins, serial quarantine for checkpoint engine tests, privacy-safe scan fixture
The v126 drift guard now compares the column SET (additive columns land at
CREATE position on fresh bootstrap but ALTER-append when migrated — ordinal
order legitimately differs) and expects the v131 checkpoint_manifest column.
Checkpoint engine tests move to a serial file (env mutation R1). The
sensitivity-scan path fixture is assembled at runtime so the banned literal
deployment path never appears in a public artifact.
