+++ test/migrations-v126.test.ts
@@ -47,10 +47,14 @@ async function captureShape() {
     is_nullable: string;
     column_default: string | null;
   }>(
+    // Name-ordered ON PURPOSE (cathedral 5): additive columns land at CREATE
+    // position on fresh bootstrap but ALTER-append on the migrated path, so
+    // ordinal_position legitimately differs — the drift guard compares the
+    // column SET + types + defaults, never physical order.
     `SELECT column_name, data_type, is_nullable, column_default
        FROM information_schema.columns
       WHERE table_name = 'session_context_state'
-      ORDER BY ordinal_position`,
+      ORDER BY column_name`,
   );
   const pk = await engine.executeRaw<{ column_name: string }>(
     `SELECT kcu.column_name
@@ -93,6 +97,7 @@ describe('migration v126 — session_context_state', () => {
     const { columns } = await captureShape();
     const byName = Object.fromEntries(columns.map(c => [c.column_name, c]));
     expect(Object.keys(byName).sort()).toEqual([
+      'checkpoint_manifest', // cathedral 5 — added by migration v131
       'client_id',
       'last_wake_at',
       'session_id',
@@ -106,6 +111,9 @@ describe('migration v126 — session_context_state', () => {
     expect(byName.session_id.data_type).toBe('text');
     expect(byName.standing_entities.data_type).toBe('jsonb');
     expect(byName.surfaced_slugs.data_type).toBe('jsonb');
+    expect(byName.checkpoint_manifest.data_type).toBe('jsonb');
+    expect(byName.checkpoint_manifest.is_nullable).toBe('NO');
+    expect(byName.checkpoint_manifest.column_default).toContain('[]');
     expect(byName.last_wake_at.data_type).toBe('timestamp with time zone');
     expect(byName.updated_at.data_type).toBe('timestamp with time zone');
     // last_wake_at is the only nullable column (no wake yet).
@@ -206,10 +214,11 @@ describe('migration v126 — session_context_state', () => {
   test('drop + rewind to v125 → runMigrations recreates the fresh-bootstrap shape', async () => {
     // Drift guard: capture the fresh-bootstrap shape (schema blob), then drop
     // the table, rewind the version ledger, and re-run migrations so v126's
-    // DDL alone recreates it. The two shapes must match exactly — a divergence
-    // means migrate.ts drifted from src/schema.sql / pglite-schema.ts.
+    // DDL + the v131 checkpoint_manifest ALTER recreate it. The two shapes
+    // must match exactly — a divergence means migrate.ts drifted from
+    // src/schema.sql / pglite-schema.ts.
     const fresh = await captureShape();
-    expect(fresh.columns.length).toBe(7);
+    expect(fresh.columns.length).toBe(8); // 7 (v126) + checkpoint_manifest (v131)
 
     await engine.executeRaw(`DROP TABLE IF EXISTS session_context_state`);
     await engine.setConfig('version', '125');
